# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail app Api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "deendipo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://retail-app.test/api",
            "description": " API Documentation Server"
        }
    ],
    "paths": {
...
        "/user/cart/view": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "View  cart ",
                "description": "View cart",
                "operationId": "viewInventoryToCart",
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "inventory_id": {
                                                        "type": "number",
                                                        "example": "28"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "4"
                                                    },
                                                    "quantity": {
                                                        "type": "number",
                                                        "example": "4"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "4"
                                                            },
                                                            "first_name": {
                                                                "type": "string",
                                                                "example": "Olamilekan"
                                                            },
                                                            "last_name": {
                                                                "type": "string",
                                                                "example": "Adeniyi"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "mailer@gmail2.com"
                                                            },
                                                            "user_type": {
                                                                "type": "string",
                                                                "example": "user"
                                                            },
                                                            "email_verified_at": {
                                                                "type": "null",
                                                                "example": "null"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "inventory": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "28"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Odun package2"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "example": "1000"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            },
                                                            "inventory_quantity": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "number",
                                                                        "example": "12"
                                                                    },
                                                                    "inventory_id": {
                                                                        "type": "number",
                                                                        "example": "28"
                                                                    },
                                                                    "quantity": {
                                                                        "type": "number",
                                                                        "example": "28"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Resouce not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}